Aranya client & daemon readmes (#21)
Initial outlines for the aranya-client and aranya-daemon readmes. A few
things:

- I included a `testing` section because I noticed in tokio they have
info for running tests. I don't feel strongly about this, but figured if
we have them we might as well include how to run them.

- I included an `example` section in the aranya-daemon readme, but since
it's a background process and I assume the code interacting with it will
be from the client, that might fit better in the aranya-client readme.

- I pulled a lot of the initial info for the aranya-daemon readme from
the top-level readme. Planning to basically pull the rest of the info
from there (for both crates)/beef it up and then link to these readmes
from the top-level one.

Closes #19

---------

Co-authored-by: Geoffrey Knopf <gknopf@spideroak.com>