contributing (#57)
Even though we have a `CONTRIBUTING.md` in the `.github` repo, it's not
easy to find.
GitHub only provides links when opening an issue or pull request as a
small link at the bottom that's barely noticeable.
The document isn't available on the repo's main landing page along with
the README and other docs and isn't easy to find after cloning the repo.
This will make it easier for users locate the correct document while
saving us the work of maintaining duplicate copies for each repo.