ci: cargo release bugfix to pull tags (#110)
The `release.yml` workflow was not pulling tags before attempting to
create new tags:

https://github.com/aranya-project/aranya/actions/runs/13775343398/job/38523245239

This change ensures that tags are pulled when checking out the repo.
This way, the `release.yml` workflow will know whether a tag exists in
the GitHub repo before attempting to create a new tag.

`continue-on-error: true` is no longer needed for the `push` step
because pulling the tags before attempting to create a new tag will
guarantee that a tag will not be created/pushed unless it is a new tag.
This means that if there is an error with pushing the tag, it is an
actual error that should cause ci to fail. Previously, it just meant
that the tag that was created locally may already exist remotely.