101/create rust example app in example dir (#129)
There was some confusion from a client about running the Rust example.
This PR adds a Rest example alongside the C example in the examples
directory. This will give the user several ways to run a Rust example.

- Question regarding https://github.com/aranya-project/aranya/pull/107.
The way we had the old template dependencies configured, we're pulling
from the crate.io releases. Which does not include the afc prefix
changes. Do we want to continue with this? Should we instead switch to
pulling in the latest `main` changes (use workspace dependencies)? I had
to back out the prefix changes.

- This recreates a previous PR.
https://github.com/aranya-project/aranya/pull/108

To test cargo generate on this PR branch run `cargo generate
aranya-project/aranya --branch
101/Create-rust-example-app-in-example-dir examples/rust --name
test-rust-example`

Closes #101, Closes #106

---------

Co-authored-by: gknopf <gknopf@spideroak.com>