aqc: finalize policy (#191)
This PR is based on the `pre-aqc-policy` branch from before any aqc
policy was added.
This was done so we can review the all the AQC policy changes on this PR
and identify any breaking changes from the original policy interface
baseline from this PR:
https://github.com/aranya-project/aranya/pull/145

Once this PR is approved, it will be merged directly into the `main`
branch.
`pre-aqc-policy` is used only for convenience so the entire AQC policy
can be reviewed rather than only the deltas from the initial baseline.

Goals for this PR:
- Ensure documentation is up-to-date
- Add missing invariants sections for various commands/queries
- Implement the `NB:` sections with correct ACL/RBAC/permissions checks
(these are borrowed from AFC).
- Generally make any changes needed to finalize the AQC policy
implementation.