ci: fix sync now test flake (#237)
The sync now test should test without adding sync peers so that when
`sync_now` is invoked, it can prove that it caused a sync to occur.

At some point, a PR introduced adding sync peers to this test, which
caused a CI flake.
Could have been one of my PRs, haven't investigated which PR introduced
it yet.