aqc: pair PSKs with cipher suites (#260)
This implementation assumes that the client only supports the cipher
suites listed in [RFC 8446]. This is a fine assumption for now, but we
might need to change it in the future if we allow users to customize the
TLS implementation.

To change this, the caller could supply the list of supported cipher
suites. Or the daemon could send the `BidiSecret`/`UniSecret` to the
client, but that's trickier since `BidiSecret` and `UniSecret` are
currently wrappers around the HPKE context, which can't reasonably be
`Serialize`/`Deserialize`.

Fixes #259

[RFC 8446]: https://datatracker.ietf.org/doc/html/rfc8446#appendix-B.4

---------

Signed-off-by: Eric Lagergren <elagergren@spideroak.com>
Co-authored-by: Jonathan Dygert <jdygert@spideroak.com>