tests: rm obsolete aranya-daemon tests (#307)
The tests in `aranya-daemon` predate the integration tests in
`aranya-client` and are now obsolete.
They currently don't offer any additional test coverage that isn't
already handled by the Aranya client integration tests that interact
with the daemon via the UDS API.
In additional to that, the tests require extra work to maintain with no
value added.
Let's remove them unless we have a valid test scenario that can't be
covered by other tests.

This will simplify some test updates that were added on the QUIC syncer
PR because there will be less test code to update after this code is
removed.