quic sync: update c api and example (#318)
Updates the following methods in the C API based on the QUIC syncer
implementation:
- `create_team()` - Generated an team PSK for the QUIC syncer
- `add_team()` - Passes in an encrypted PSK seed for the QUIC syncer
- `remove_team()` - Tests the API method in the C example now that's
it's implemented by https://github.com/aranya-project/aranya/pull/302

Methods added to the C API on this PR:
- `get_psk_seed_encrypted()` - Gets a PSK seed encrypted for a specific
device

The team config has been updated to include the QUIC syncer PSK seed
(will be changed to an encrypted PSK seed soon once that's unblocked).

When creating a team with `create_team()`, a PSK seed is generated to be
used by all QUIC syncer client/servers when syncing for that team.
The PSK seed should be encrypted with a device's public encryption key
before it is sent to the device to be used when adding a team with
`add_team()`.

Uses the updated/new methods in the C example app.

Solves:
https://github.com/aranya-project/aranya/issues/317
https://github.com/aranya-project/aranya/issues/167

---------

Co-authored-by: Jonathan Dygert <jdygert@spideroak.com>