ci: don't cancel concurrent workflows on main branch (#417)
Concurrency groups are used by our CI pipeline to cancel older workflows
in favor of running new ones for new commits.
This reduces CI usage by cancelling workflows that are no longer needed.

The problem with this is that while this strategy works well for
development branches/PRs, it is not a good approach for `main`,
`master`, `release`, `production`, etc. branches. These types of
branches must guarantee each commit builds, passes tests, completes
upload of release artifacts, or completes deployment of software to a
production server. Cancelling such jobs could have unintended
consequences such as releasing software with partial artifacts or
partially deploying software to a server which results in a broken
deployment.

The solution to this is to skip concurrency group cancellation if a
branch matches a certain pattern. The only branch we're concerned about
at this time is `main`, although this could extend to other branches in
the future.

Relevant docs:

https://docs.github.com/en/actions/how-tos/writing-workflows/choosing-when-your-workflow-runs/control-the-concurrency-of-workflows-and-jobs#example-only-cancel-in-progress-jobs-on-specific-branches