c-example: fix crash on errors in example (#505)
This should fix the crashes that occurred when we had other errors.

We were "cleaning up" uninitialized objects when an error occurred. We
now basically just exit on an error. There isn't really a need to do the
cleanup here on errors. It's probably worth having in a more realistic
example but this example is basically just a test script.