custom roles: remove label assignment to roles from policy (#561)
Remove support for assigning labels to roles from the policy.
This simplifies AFC channel deletion handling because we don't need to
compare the set of labels assigned to a device's role to the set of
labels assigned directly to the device when determining whether a
channel can be deleted.
If a label is not assigned to a device, then any AFC channels the device
is participating in can be deleted.