policy: use enums for perms everywhere (#656)
We are currently using enums for permissions in the Aranya client Rust
and C APIs.
These are converted into a `Text` string at the daemon IPC API and
passed as `string` to the policy interface.
Internally, the policy converts the strings to an enum again for
commands/facts, then converts back to a string before returning the
permission via an effect.

There is no reason to perform all these type conversions when
permissions can be represented as an enum everywhere.